refactor: create network message system
