+++ rustyhack_server/src/game/player_updates.rs
@@ -211,95 +211,6 @@ fn send_player_joined_response(player: &Player, sender: &Sender<Packet>) {
     );
 }
 
-pub(crate) fn send_player_position_updates(world: &mut World, sender: &Sender<Packet>) {
-    let mut query = <(&PlayerDetails, &mut Position)>::query();
-    for (player_details, position) in query.iter_mut(world) {
-        if position.update_available && player_details.currently_online {
-            debug!(
-                "Sending player position update for: {}",
-                &player_details.player_name
-            );
-            let response = serialize(&ServerMessage::UpdatePosition(position.clone()))
-                .unwrap_or_else(|err| {
-                    error!(
-                        "Failed to serialize player position: {:?}, error: {}",
-                        &position, err
-                    );
-                    process::exit(1);
-                });
-            rustyhack_lib::message_handler::send_packet(
-                Packet::unreliable_sequenced(
-                    player_details.client_addr.parse().unwrap(),
-                    response,
-                    Some(20),
-                ),
-                sender,
-            );
-            position.update_available = false;
-        }
-    }
-    debug!("Finished sending player position updates.");
-}
-
-pub(crate) fn send_player_stats_updates(world: &mut World, sender: &Sender<Packet>) {
-    let mut query = <(&PlayerDetails, &mut Stats)>::query();
-    for (player_details, stats) in query.iter_mut(world) {
-        if stats.update_available && player_details.currently_online {
-            debug!(
-                "Sending player stats update for: {}",
-                &player_details.player_name
-            );
-            let response = serialize(&ServerMessage::UpdateStats(*stats)).unwrap_or_else(|err| {
-                error!(
-                    "Failed to serialize player stats: {:?}, error: {}",
-                    &stats, err
-                );
-                process::exit(1);
-            });
-            rustyhack_lib::message_handler::send_packet(
-                Packet::unreliable_sequenced(
-                    player_details.client_addr.parse().unwrap(),
-                    response,
-                    Some(21),
-                ),
-                sender,
-            );
-            stats.update_available = false;
-        }
-    }
-    debug!("Finished sending player stats updates.");
-}
-
-pub(crate) fn send_player_inventory_updates(world: &mut World, sender: &Sender<Packet>) {
-    let mut query = <(&PlayerDetails, &mut Inventory)>::query();
-    for (player_details, inventory) in query.iter_mut(world) {
-        if inventory.update_available && player_details.currently_online {
-            debug!(
-                "Sending player inventory update for: {}",
-                &player_details.player_name
-            );
-            let response = serialize(&ServerMessage::UpdateInventory(inventory.clone()))
-                .unwrap_or_else(|err| {
-                    error!(
-                        "Failed to serialize player inventory: {:?}, error: {}",
-                        &inventory, err
-                    );
-                    process::exit(1);
-                });
-            rustyhack_lib::message_handler::send_packet(
-                Packet::unreliable_sequenced(
-                    player_details.client_addr.parse().unwrap(),
-                    response,
-                    Some(24),
-                ),
-                sender,
-            );
-            inventory.update_available = false;
-        }
-    }
-    debug!("Finished sending player stats updates.");
-}
-
 pub(crate) fn send_other_entities_updates(world: &World, sender: &Sender<Packet>) {
     let mut position_updates: HashMap<String, Position> = HashMap::new();
     let mut display_details: HashMap<String, DisplayDetails> = HashMap::new();
